backend/requirements.txt
transformers
torch
pillow
flask
spacy
sentence-transformers
